msm: kgsl: Correctly identify page faulting process
Currently, we rely on the global current_context entry in memstore
to identify the faulting context in the pagefault handler. This entry
could be wrong during preemption from higher priority RB to a lower
one. This happens because the global entry in memstore is updated
using inline PM4 packets in the RB. So after resuming a preempted RB,
the global current_context value will be wrong until there is a
context switch in the new RB.

We can avoid identifying the faulting context. It is enough to identify
the faulting process and we can do that from pagetable base address
from the TTBR0 register.

Change-Id: Ib3da7b55f8245350a4430ff97ec7c34804a41b07
Signed-off-by: Akhil P Oommen <akhilpo@codeaurora.org>
Signed-off-by: Sravankumar bijili <csbijil@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>